bigquery/projects/import_from_mysql/batch_and_stream/requirements.txt
pymysql
google-cloud-bigquery
google-cloud-storage
cloud-sql-python-connector[pymysql]
